ClusterMarker.js added
